chore: 🐋 Tile, ExpandTile 폴더 분리
index.ts 추가, Tile.styles 분리
