+++ src/components/atoms/Tile/Tile/index.ts
@@ -1,0 +1,10 @@
+import Tile from "./Tile";
+import { tileVariants } from "./Tile.styles";
+
+export type { TileDefault, TileAlterAs, TileProps } from "./Tile.types";
+
+// style export
+export { tileVariants };
+
+// component export
+export default Tile;
